moving and refactoring necessary files to new module and removing the jumps module
