+++ features/jump/src/main/java/com/fredprojects/features/jump/data/mapper/JDMapper.kt
@@ -1,0 +1,15 @@
+package com.fredprojects.features.jump.data.mapper
+
+import com.fredprojects.core.database.entities.JDEntity
+import com.fredprojects.features.jump.domain.models.JumpData
+
+internal fun JumpData.toEntity() = JDEntity(
+    count = count,
+    date = date,
+    id = id
+)
+internal fun JDEntity.toDomain() = JumpData(
+    count = count,
+    date = date,
+    id = id
+)
