+++ features/jump/src/main/java/com/fredprojects/features/jump/domain/usecases/JDUseCases.kt
@@ -1,0 +1,9 @@
+package com.fredprojects.features.jump.domain.usecases
+
+import com.fredprojects.features.jump.domain.usecases.crud.*
+
+data class JDUseCases(
+    val getData: GetJDUseCase,
+    val upsert: UpsertJDUseCase,
+    val delete: DeleteJDUseCase
+)
