+++ features/jump/src/main/java/com/fredprojects/features/jump/domain/utils/SortType.kt
@@ -1,0 +1,6 @@
+package com.fredprojects.features.jump.domain.utils
+
+sealed class SortType {
+    data object Ascending : SortType()
+    data object Descending : SortType()
+}
