--- features/jumps/data/.gitignore
@@ -1,1 +1,0 @@
-/build
