--- features/jumps/domain/build.gradle.kts
@@ -1,13 +1,0 @@
-plugins {
-    id("java-library")
-    alias(libs.plugins.kotlin.jvm)
-}
-
-java {
-    sourceCompatibility = JavaVersion.VERSION_17
-    targetCompatibility = JavaVersion.VERSION_17
-}
-
-dependencies {
-    implementation(libs.kotlinx.coroutines.core)
-}
