--- features/jumps/domain/src/main/java/com/fredprojects/features/jumps/domain/useCases/JDUseCases.kt
@@ -1,9 +1,0 @@
-package com.fredprojects.features.jumps.domain.useCases
-
-import com.fredprojects.features.jumps.domain.useCases.crud.*
-
-data class JDUseCases(
-    val getData: GetJDUseCase,
-    val upsert: UpsertJDUseCase,
-    val delete: DeleteJDUseCase
-)
