--- features/jumps/domain/src/main/java/com/fredprojects/features/jumps/domain/utils/JDVerification.kt
@@ -1,17 +1,0 @@
-package com.fredprojects.features.jumps.domain.utils
-
-import com.fredprojects.features.jumps.domain.models.JumpData
-import java.time.LocalDate
-import java.time.format.DateTimeFormatter
-import java.time.format.DateTimeParseException
-
-internal fun JumpData.isCountCorrect() = count > 0
-internal fun JumpData.isDateCorrect() = try {
-    if (date.contains(Regex("""^\d{4}-\d{2}-\d{2}$"""))) {
-        val localDate = LocalDate.parse(date, DateTimeFormatter.ofPattern("yyyy-MM-dd"))
-        localDate <= LocalDate.now()
-    } else false
-} catch (e: DateTimeParseException) {
-    false
-}
-internal fun JumpData.isValid() = isCountCorrect() || isDateCorrect()
