--- features/jumps/domain/src/main/java/com/fredprojects/features/jumps/domain/utils/JumpStatus.kt
@@ -1,17 +1,0 @@
-package com.fredprojects.features.jumps.domain.utils
-
-/**
- * JumpStatus represents the status of the JumpData
- * @property SUCCESS if the JumpData was successfully inserted or updated
- * @property INCORRECT_DATE if the date is incorrect
- * @property INCORRECT_COUNT if the count is incorrect
- * @property INCORRECT_DATA if the data is incorrect
- * @property NOTHING if there is no status
- */
-enum class JumpStatus {
-    SUCCESS,
-    INCORRECT_DATE,
-    INCORRECT_COUNT,
-    INCORRECT_DATA,
-    NOTHING
-}
