--- features/jumps/presentation/src/main/java/com/fredprojects/features/jumps/presentation/models/JDPModel.kt
@@ -1,17 +1,0 @@
-package com.fredprojects.features.jumps.presentation.models
-
-import androidx.compose.runtime.Stable
-import java.time.LocalDate
-
-/**
- * JumpDataPresentationModel is used to store data of the jumps.
- * @param count the number of jumps
- * @param date the date of the training
- * @param id the identifier of the record
- */
-@Stable
-data class JDPModel(
-    val count: Int?,
-    val date: String = LocalDate.now().toString(),
-    val id: Int? = null
-)
