+++ settings.gradle.kts
@@ -33,12 +33,9 @@ include(":features:pws:domain")
 include(":features:pws:data")
 include(":features:pws:presentation")
 
-include(":features:jumps:domain")
-include(":features:jumps:data")
-include(":features:jumps:presentation")
-
 include(":features:auth:domain")
 include(":features:auth:data")
 include(":features:auth:presentation")
+
 include(":features:math")
 include(":features:jump")
